:dizzy: Break out the Service Worker into sub-components
Easier to maintain
