+++ src/sw.js
@@ -13,156 +13,14 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
-/* global Set, Promise, importScripts, languages  */
 import { CacheableResponsePlugin } from 'workbox-cacheable-response';
 import { CacheFirst, StaleWhileRevalidate } from 'workbox-strategies';
 import { ExpirationPlugin } from 'workbox-expiration';
-import { matchPrecache, precacheAndRoute } from 'workbox-precaching';
+import { precacheAndRoute } from 'workbox-precaching';
 import { registerRoute, setCatchHandler } from 'workbox-routing';
-import { preferences } from './js/lib/db.js';
-
-// TODO: You'll probably want to move from importScripts() to an ES import
-// for these scripts, since you're already running this file through Rollup.
-importScripts('/js/languages.js');
-
-/**
- * Normalizes requests that end with a backslash (/) to request the same resource
- * From https://github.com/GoogleChrome/workbox/issues/2217#issuecomment-528023769
- *
- * @param {object} param0 - An object containing a Workbox request object
- * @return {request} Either the original request or a normalized request
- */
-async function normalizeIfNeeded({ request }) {
-  // Clean out query parameters, and add a trailing '/' if missing.
-  const url = new URL(request.url);
-  let cleanUrl = url.origin + url.pathname;
-  if (!cleanUrl.endsWith('/')) {
-    cleanUrl += '/';
-  }
-  return new Request(cleanUrl);
-}
-
-const includesRegExp = /(<!--\s*#include\s*virtual=['|"]\S*['|"]-->)/gm;
-const includesFileRegExp = /<!--\s*#include\s*virtual=['|"](\S*)['|"]-->/gm;
-const endIncludeRegExp = /<!--\s*#endinclude\s*-->/gm;
-const endIncludeWithLeadingRegExp = /[\s\S]*<!--\s*#endinclude\s*-->/gm;
-
-/**
- *
- * @param {response} param0 - The response from the cache
- */
-async function serviceWorkerSideInclude({ cachedResponse }) {
-  const content = await cachedResponse.text();
-  const matches = [...new Set(content.match(includesRegExp))];
-  const neededIncludes = await Promise.all(
-    matches
-      .map(i => i.split(includesFileRegExp)[1])
-      .map(async key => {
-        const cachedInclude = await matchPrecache(key);
-        return cachedInclude.text();
-      }),
-  );
-
-  const includes = {};
-
-  matches.forEach((include, i) => (includes[include] = neededIncludes[i]));
-
-  const rebuild = content
-    .split(includesRegExp)
-    .map(i => {
-      if (matches.includes(i)) {
-        return includes[i];
-      }
-
-      return i;
-    })
-    .join('');
-
-  return new Response(rebuild, { headers: cachedResponse.headers });
-}
-
-/**
- *
- * @param {response} param0 - The response that will update the cache
- */
-async function swsiSideCleanup({ response }) {
-  const content = await response.text();
-
-  const matches = content.match(includesRegExp);
-  // const neededIncludes = [...new Set(matches)].map(i => i.split(includesFileRegExp)[1]);
-  let open = 0;
-  const rebuild = content
-    .split(includesRegExp)
-    .map(i => {
-      // If the current item is the include and it's not included from within
-      if (i === matches[0]) {
-        matches.shift();
-        open++;
-        if (open > 1) return '';
-        return i;
-      }
-
-      const endIncludeSplit = i.split(endIncludeWithLeadingRegExp);
-      if (endIncludeSplit.length === 1 && open !== 0) {
-        return '';
-      }
-
-      const count = i.match(endIncludeRegExp);
-
-      open = open - (count ? count.length : 0);
-
-      return endIncludeSplit.join('');
-    })
-    .join('');
-
-  return new Response(rebuild, { headers: response.headers });
-}
-
-const navigationNormalizationPlugin = {
-  cacheKeyWillBeUsed: normalizeIfNeeded,
-  requestWillFetch: normalizeIfNeeded,
-  cachedResponseWillBeUsed: serviceWorkerSideInclude,
-  cacheWillUpdate: swsiSideCleanup,
-};
+import { htmlHandler } from './js/sw/html-strategy';
 
 precacheAndRoute(self.__WB_MANIFEST);
-
-// HTML caching strategy
-const htmlStrategy = new StaleWhileRevalidate({
-  cacheName: 'pages-cache',
-  plugins: [
-    navigationNormalizationPlugin,
-    new CacheableResponsePlugin({
-      statuses: [200, 301, 404],
-    }),
-  ],
-});
-
-/**
- * HTML Handler
- * @param {object} param0 - Workbox Handler
- * @return {object} - Either a redirection response or a Workbox response
- */
-async function htmlHandler({ event }) {
-  const lang = await preferences.get('lang');
-  const { request } = event;
-
-  const urlLang = request.url.replace(self.location.origin, '').split('/');
-  const currentLang = urlLang[1];
-  const isALanguage = languages.includes(currentLang);
-  const isRightLanguage = currentLang !== lang;
-  const shouldRefresh = new URL(request.url).searchParams.get('locale_fallback') !== 'true';
-
-  if (isALanguage && isRightLanguage && shouldRefresh) {
-    console.log('⇒ Redirecting');
-    urlLang[1] = lang;
-    const redirectURL = `${self.location.origin}${urlLang.join('/')}`;
-    return Response.redirect(redirectURL, 302);
-  }
-
-  return htmlStrategy.handle({ event, request });
-}
 
 // Handle navigation requests with htmlHandler.
 registerRoute(({ request }) => request.mode === 'navigate', htmlHandler);
